api/requirements.txt
Flask-Cors
Flask
Flask-RESTful
uWSGI
python-dotenv
requests